Undo changes to remove FirebaseApp split
