+++ test/resources/mocks.ts
@@ -26,8 +26,7 @@ import * as jwt from 'jsonwebtoken';
 
 import { FirebaseNamespace } from '../../src/firebase-namespace';
 import { FirebaseServiceInterface } from '../../src/firebase-service';
-import { FirebaseApp } from '../../src/firebase-app';
-import { FirebaseAppOptions } from '../../src/firebase-app-internal';
+import { FirebaseApp, FirebaseAppOptions } from '../../src/firebase-app';
 import { Credential, GoogleOAuthAccessToken, ServiceAccountCredential } from '../../src/auth/credential';
 
 const ALGORITHM = 'RS256';
