+++ test/unit/firebase-app.spec.ts
@@ -27,8 +27,7 @@ import * as mocks from '../resources/mocks';
 
 import { GoogleOAuthAccessToken, ServiceAccountCredential } from '../../src/auth/credential';
 import { FirebaseServiceInterface } from '../../src/firebase-service';
-import { FirebaseApp } from '../../src/firebase-app';
-import { FirebaseAccessToken } from '../../src/firebase-app-internal';
+import { FirebaseApp, FirebaseAccessToken } from '../../src/firebase-app';
 import { FirebaseNamespace, FirebaseNamespaceInternals, FIREBASE_CONFIG_VAR } from '../../src/firebase-namespace';
 
 import { Auth } from '../../src/auth/auth';
