+++ test/unit/utils.ts
@@ -20,8 +20,7 @@ import * as sinon from 'sinon';
 import * as mocks from '../resources/mocks';
 
 import { FirebaseNamespace } from '../../src/firebase-namespace';
-import { FirebaseApp } from '../../src/firebase-app';
-import { FirebaseAppOptions, FirebaseAppInternals, FirebaseAccessToken } from '../../src/firebase-app-internal';
+import { FirebaseApp, FirebaseAppOptions, FirebaseAppInternals, FirebaseAccessToken } from '../../src/firebase-app';
 import { HttpError, HttpResponse } from '../../src/utils/api-request';
 
 /**
